+++ src/pages/TimerPage.jsx
@@ -1,10 +1,12 @@
-import { useState, useRef, useEffect, useReducer } from 'react';
+import { useRef, useEffect, useReducer } from 'react';
 import styled from 'styled-components';
 
 import { TimerButton, SmallTimerbutton } from '../components/TimerButton';
 import TimerPannel from '../components/TimerPannel';
 import Modal from '../components/Modal';
 import basicAlarm from '../asset/basicAlarm.mp3';
+
+import useTimer from '../hooks/useTimer';
 
 const ContentBox = styled.div`
   width: 100%;
@@ -30,44 +32,54 @@ const ModalMessage = styled.div`
 `;
 
 const TimerPage = () => {
-  const [time, setTime] = useState(0);
-  const [isRunning, setIsRunning] = useState(false);
-  const timerId = useRef(null);
-  const initialTime = useRef(0);
-  const audioRef = useRef(null);
+  // const [time, setTime] = useState(0);
+  // const [isRunning, setIsRunning] = useState(false);
 
+  const {
+    time,
+    timerId,
+    startTimer,
+    stopTimer,
+    resetTimer,
+    modifyTimer,
+    isRunning,
+    setTime,
+    initialTime,
+    setIsRunning,
+  } = useTimer();
+  const audioRef = useRef(null);
   const [isOpen, toggleModal] = useReducer((modalOpen) => {
     modalOpen && audioRef.current.pause();
     return !modalOpen;
   }, false);
 
-  const startTimer = () => {
-    setIsRunning(true);
-    timerId.current = setInterval(() => {
-      setTime((prev) => prev - 1);
-    }, 1000);
-  };
+  // const startTimer = () => {
+  //   setIsRunning(true);
+  //   timerId.current = setInterval(() => {
+  //     setTime((prev) => prev - 1);
+  //   }, 1000);
+  // };
 
-  const stopTimer = () => {
-    clearInterval(timerId.current);
-    setIsRunning(false);
-  };
+  // const stopTimer = () => {
+  //   clearInterval(timerId.current);
+  //   setIsRunning(false);
+  // };
 
-  const modifyTimer = () => {
-    const newMinute = prompt('몇 분를 셀 것인지 입력해주세요');
-    const newSecond = prompt('몇 초를 셀 것인지 입력해주세요');
+  // const modifyTimer = () => {
+  //   const newMinute = prompt('몇 분를 셀 것인지 입력해주세요');
+  //   const newSecond = prompt('몇 초를 셀 것인지 입력해주세요');
 
-    const calcSecond = Number(newMinute) * 60 + Number(newSecond);
-    setTime(calcSecond);
+  //   const calcSecond = Number(newMinute) * 60 + Number(newSecond);
+  //   setTime(calcSecond);
 
-    initialTime.current = calcSecond;
-  };
+  //   initialTime.current = calcSecond;
+  // };
 
-  const resetTimer = () => {
-    setTime(initialTime.current);
-    clearInterval(timerId.current);
-    setIsRunning(false);
-  };
+  // const resetTimer = () => {
+  //   setTime(initialTime.current);
+  //   clearInterval(timerId.current);
+  //   setIsRunning(false);
+  // };
 
   const onClickRestart = () => {
     toggleModal();
@@ -102,7 +114,7 @@ const TimerPage = () => {
       clearInterval(timerId.current);
       setIsRunning(false);
     }
-  }, [time, isRunning]);
+  }, [time, isRunning, setIsRunning, timerId]);
 
   return (
     <ContentBox>
